refactor(racing): move to model package
