+++ src/main/java/racing/model/CarCondition.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.model;
 
 import java.util.Random;
