+++ src/main/java/racing/model/CarName.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.model;
 
 import org.apache.commons.lang3.StringUtils;
