+++ src/main/java/racing/model/Position.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.model;
 
 public class Position {
