+++ src/test/java/racing/CarNameTest.java
@@ -1,9 +1,9 @@
 package racing;
 
 import org.junit.jupiter.api.Test;
+import racing.model.CarName;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class CarNameTest {
