+++ src/test/java/racing/CarTest.java
@@ -2,6 +2,9 @@ package racing;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import racing.model.Car;
+import racing.model.CarName;
+import racing.model.Position;
 
 import static org.assertj.core.api.Assertions.assertThat;
