+++ src/test/java/racing/PositionTest.java
@@ -2,6 +2,7 @@ package racing;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import racing.model.Position;
 
 import static org.assertj.core.api.Assertions.assertThat;
